fix: upgrade to alpha-6
and fix typescript errors caused by divergence in types between Fed 1 and Fed 2 query plans
